Finished

+++ Unit Testing/UnitTesting/Game.h
@@ -10,11 +10,11 @@ public:
 
 	Game() {};
 	~Game() {};
-	std::set<float> generateNumbers(int t_amount);
+	std::set<int> generateNumbers(int t_amount);
 
 private:
 
 	int m_randomNum = 0;
-	std::set<float> m_numberList;
+	std::set<int> m_numberList;
 };
 #endif
